requirements.txt
accelerate==1.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
diffusers==0.32.2
filelock==3.17.0
fsspec==2025.2.0
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pillow==11.1.0
psutil==6.1.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
setuptools==75.8.0
sympy==1.13.1
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.2
typing_extensions==4.12.2
urllib3==2.3.0
zipp==3.21.0
